Space Invaders contribution: Hanzi Space Invaders
Hanzi Space Invaders is a part of several games I had made with p5js and React with the aim to make learning Hanzi fun and interactive! The Coding Train Space Invaders challenge was a great inspiration for me to make this game, and it was built on the baseline of that challenge.